renamed grid_resilience.py to power_outage.py
